[kernel: scheduler] start implementing a scheduler
this commit adds a half-working scheduler to the kernel.
the scheduler will allow us to do multiple tasks at once by spreading them across all threads. at the moment, this throws a general protection fault when changing the task to a new ome, i will be investigating this soon :P